+++ src/label/bb_labelling.cpp
@@ -22,7 +22,7 @@ warthog::label::bb_labelling::print(
         std::ostream& out, uint32_t first_id, uint32_t last_id)
 {
     if(first_id > labels_->size()) { return; }
-    if(last_id > labels_->size()) { last_id = labels_->size(); }
+    if(last_id > labels_->size()) { last_id = (uint32_t)labels_->size(); }
 
     out << "# Before printing labels are sorted by tail index. \n"
         << "# Ties are broken using the order edges appear in the file \n"
