+++ src/util/experiment.cpp
@@ -12,6 +12,6 @@ warthog::experiment::print(std::ostream& out)
 	out << this->starty()<<"\t";
 	out << this->goalx()<<"\t";
 	out << this->goaly()<<"\t";
-	out << std::fixed << std::setprecision(this->precision());
+	out << std::fixed << std::setprecision((int)this->precision());
 	out << this->distance();
 }
